requirements.txt
asgiref==3.5.2
bcrypt==3.2.0
blinker==1.4
certifi==2020.6.20
cffi==1.15.0
chardet==4.0.0
charset-normalizer==2.0.12
click==8.0.3
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
cryptography==3.4.8
distlib==0.3.4
distro==1.7.0
Django==3.2.12
django-ckeditor==6.4.2
django-colorfield==0.7.0
django-colorful==1.3
django-countries==7.3.2
django-filter==21.1
django-js-asset==2.0.0
django-phonenumber-field==6.1.0
django-rest-swagger==2.2.0
djangorestframework==3.13.1
drf-yasg==1.20.0
fasteners==0.14.1
filelock==3.7.1
future==0.18.2
httplib2==0.20.2
idna==3.3
importlib-metadata==4.6.4
inflection==0.5.1
itypes==1.2.0
jeepney==0.7.1
Jinja2==3.1.2
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
oauthlib==3.2.0
olefile==0.46
openapi-codec==1.3.2
packaging==21.3
paramiko==2.9.3
pexpect==4.8.0
phonenumbers==8.12.49
Pillow==9.0.1
platformdirs==2.5.2
protobuf==3.12.4
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pycparser==2.21
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
python-dateutil==2.8.1
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
SecretStorage==3.3.1
simplejson==3.17.6
six==1.16.0
sqlparse==0.4.2
ssh-import-id==5.11
typing_extensions==4.2.0
uritemplate==4.1.1
urllib3==1.26.5
virtualenv==20.14.1
wadllib==1.3.6
xdg==5.0.0
zipp==1.0.0
